changes requested applied
