--- src/config/auth,json
@@ -1,3 +1,0 @@
-{
-    "secret": "d41d8cd98f00b204e9800998ecf8427e" 
-}
