+++ src/routes/authRoutes.js
@@ -1,12 +1,53 @@
 const express = require('express');
 
+const router = express.Router();
+
 const jsonwebtoken = require('jsonwebtoken');
 
-const router = express.Router();
+const User = require('../models/user');
+
+router.get('/', (req, res) => {
+    res.json({
+        message: 'Authentication!'
+    });
+});
 
 const authConfig = {
     "secret": "d41d8cd98f00b204e9800998ecf8427e" 
 };
+
+router.post('/login', async(req, res) => {    
+    const {email, password} = req.body;
+
+    //const user = await User.findOne({ email }).select('+password');
+    const user = {
+        "email": "asda@asd.com",
+        "password": "asdz"
+    };
+
+    if(!user){
+        
+        return res.status(400).send({Error: 'User not found'});
+
+    }
+
+    if(!await password == user.password){
+        
+        return res.status(400).send({ Error: 'Incorrect password'});
+
+    }
+
+    user.password = undefined;
+
+    const token = jsonwebtoken.sign({id: user.id}, authConfig.secret,{
+        expiresIn: 86400,
+    }); 
+
+    const aToken = "Bearer "+token;
+
+    res.header("authToken", aToken);
+    res.redirect('/main');
+});
 
 router.get('/logout', async(req, res) => { 
     const sessiontoken = req.headers.sessiontoken;
@@ -41,4 +82,4 @@ router.get('/logout', async(req, res) => {
     res.redirect('/login');
 });
 
-module.exports = app => app.use('/', router);
+module.exports = router;
